Laptop-2/requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pyrebase4==4.7.1
Pillow==10.0.0
python-dateutil==2.8.2
requests
